Make EF test namespaces consistent
This seems to resolve some flakey tests in the VSCode test runner
