+++ test/EntityFramework.IntegrationTests/DatabaseProviderFixture.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 
-namespace Tests;
+namespace EntityFramework.IntegrationTests;
 
 /// <summary>
 /// xUnit ClassFixture for creating and deleting integration test databases.
