+++ test/EntityFramework.IntegrationTests/IntegrationTest.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Xunit;
 
-namespace Tests;
+namespace EntityFramework.IntegrationTests;
 
 /// <summary>
 /// Base class for integration tests, responsible for initializing test database providers & an xUnit class fixture
@@ -37,19 +37,19 @@ public class IntegrationTest<TClass, TDbContext, TStoreOption> : IClassFixture<D
 
             TestDatabaseProviders = new TheoryData<DbContextOptions<TDbContext>>
             {
-                //DatabaseProviderBuilder.BuildInMemory<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions),
+                // DatabaseProviderBuilder.BuildInMemory<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions),
                 DatabaseProviderBuilder.BuildSqlite<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions),
-                //DatabaseProviderBuilder.BuildLocalDb<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions)
+                DatabaseProviderBuilder.BuildLocalDb<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions)
             };
         }
         else
         {
             TestDatabaseProviders = new TheoryData<DbContextOptions<TDbContext>>
             {
-                //DatabaseProviderBuilder.BuildInMemory<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions),
-                DatabaseProviderBuilder.BuildSqlite<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions)
+                // DatabaseProviderBuilder.BuildInMemory<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions),
+                // DatabaseProviderBuilder.BuildSqlite<TDbContext, TStoreOption>(typeof(TClass).Name, StoreOptions)
             };
-            Console.WriteLine("Skipping DB integration tests on non-Windows");
+            // Console.WriteLine("Skipping DB integration tests on non-Windows");
         }
     }
