+++ test/EntityFramework.Storage.IntegrationTests/DatabaseProviderFixture.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 
-namespace IntegrationTests;
+namespace EntityFramework.Storage.IntegrationTests;
 
 /// <summary>
 /// xUnit ClassFixture for creating and deleting integration test databases.
