+++ test/EntityFramework.Storage.IntegrationTests/Stores/ResourceStoreTests.cs
@@ -16,7 +16,7 @@ using IdentityModel;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
 
-namespace IntegrationTests.Stores;
+namespace EntityFramework.Storage.IntegrationTests.Stores;
 
 public class ScopeStoreTests : IntegrationTest<ScopeStoreTests, ConfigurationDbContext, ConfigurationStoreOptions>
 {
