+++ test/EntityFramework.Storage.UnitTests/Mappers/ApiResourceMappersTests.cs
@@ -9,7 +9,7 @@ using Xunit;
 using Models = Duende.IdentityServer.Models;
 using Entities = Duende.IdentityServer.EntityFramework.Entities;
 
-namespace UnitTests.Mappers;
+namespace EntityFramework.Storage.UnitTests.Mappers;
 
 public class ApiResourceMappersTests
 {
