+++ test/EntityFramework.Storage.UnitTests/Mappers/PersistedGrantMappersTests.cs
@@ -8,7 +8,7 @@ using Xunit;
 using Models = Duende.IdentityServer.Models;
 using Entities = Duende.IdentityServer.EntityFramework.Entities;
 
-namespace UnitTests.Mappers;
+namespace EntityFramework.Storage.UnitTests.Mappers;
 
 public class PersistedGrantMappersTests
 {
